requirements.txt
black==19.10b0
celery==4.4.2
Django==3.0.6
django-autoslug==1.9.7
djangorestframework==3.11.0
flake8==3.8.2
psycopg2-binary==2.8.5
redis==3.5.2
djoser==2.0.3
